# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API with Stripe Payments",
        "description": "A comprehensive e-commerce API with Stripe payment integration, real-time WebSocket features, and comprehensive order management",
        "contact": {
            "email": "admin@ecommerce-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "http://e-commerce-api-new.test/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Get dashboard statistics",
                "description": "Retrieve key metrics and statistics for the admin dashboard including counts, revenue, and recent orders",
                "operationId": "3e557e4f514f829bac52cb3641f404d9",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "products_count": {
                                            "description": "Total number of products",
                                            "type": "integer",
                                            "example": 125
                                        },
                                        "categories_count": {
                                            "description": "Total number of categories",
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "orders_count": {
                                            "description": "Total number of non-canceled orders",
                                            "type": "integer",
                                            "example": 89
                                        },
                                        "pending_orders": {
                                            "description": "Number of pending orders",
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "delivered_orders": {
                                            "description": "Number of delivered orders",
                                            "type": "integer",
                                            "example": 45
                                        },
                                        "total_revenue": {
                                            "description": "Total revenue from paid orders",
                                            "type": "string",
                                            "example": "15420.75"
                                        },
                                        "latest_orders": {
                                            "description": "5 most recent orders",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 22
                                                    },
                                                    "total_amount": {
                                                        "type": "string",
                                                        "example": "299.99"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "payment_status": {
                                                        "type": "string",
                                                        "example": "unpaid"
                                                    },
                                                    "payment_method": {
                                                        "type": "string",
                                                        "example": "stripe"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-10-31T14:30:00Z"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Jane Smith"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "jane@example.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}